+++ src/pages/@owner/food-truck-form/@section/category-section/NeedElectricity.tsx
@@ -1,4 +1,4 @@
-import FormLayout from '@pages/@owner/food-truck-form/components/FormLayout';
+import FormLayout from '@components/layout/form-layout/FormLayout';
 import Button from '@ui/button/Button';
 import { NEED_ELECTRICITY } from '@constant/need-electricity';
 import { useCategories } from '@pages/@owner/food-truck-form/hooks/use-categories';
